+++ sdk/src/org.graalvm.launcher/src/org/graalvm/launcher/AbstractLanguageLauncher.java
@@ -182,7 +182,7 @@ public abstract class AbstractLanguageLauncher extends Launcher {
             if (languageName == null || languageName.length() == 0) {
                 languageName = languageId;
             }
-            languageImplementationName = "Graal " + languageName;
+            languageImplementationName = languageName;
         }
         String engineImplementationName = engine.getImplementationName();
         if (isAOT()) {
@@ -190,7 +190,13 @@ public abstract class AbstractLanguageLauncher extends Launcher {
         } else {
             engineImplementationName += " JVM";
         }
-        System.out.println(String.format("%s %s (%s %s)", languageImplementationName, language.getVersion(), engineImplementationName, engine.getVersion()));
+        String languageVersion = language.getVersion();
+        if (languageVersion.equals(engine.getVersion())) {
+            languageVersion = "";
+        } else {
+            languageVersion += " ";
+        }
+        System.out.println(String.format("%s %s(%s %s)", languageImplementationName, languageVersion, engineImplementationName, engine.getVersion()));
     }
 
     protected void runVersionAction(VersionAction action, Engine engine) {